RIC-973 Policy status notification handling - initial rollup

+++ pkg/restapi/operations/a1_mediator/a1_controller_delete_policy_type_parameters.go
@@ -36,7 +36,8 @@ import (
 )
 
 // NewA1ControllerDeletePolicyTypeParams creates a new A1ControllerDeletePolicyTypeParams object
-// no default values defined in spec.
+//
+// There are no default values defined in the spec.
 func NewA1ControllerDeletePolicyTypeParams() A1ControllerDeletePolicyTypeParams {
 
 	return A1ControllerDeletePolicyTypeParams{}
@@ -74,7 +75,6 @@ func (o *A1ControllerDeletePolicyTypeParams) BindRequest(r *http.Request, route 
 	if err := o.bindPolicyTypeID(rPolicyTypeID, rhkPolicyTypeID, route.Formats); err != nil {
 		res = append(res, err)
 	}
-
 	if len(res) > 0 {
 		return errors.CompositeValidationError(res...)
 	}
@@ -107,11 +107,11 @@ func (o *A1ControllerDeletePolicyTypeParams) bindPolicyTypeID(rawData []string, 
 // validatePolicyTypeID carries on validations for parameter PolicyTypeID
 func (o *A1ControllerDeletePolicyTypeParams) validatePolicyTypeID(formats strfmt.Registry) error {
 
-	if err := validate.MinimumInt("policy_type_id", "path", int64(o.PolicyTypeID), 1, false); err != nil {
+	if err := validate.MinimumInt("policy_type_id", "path", o.PolicyTypeID, 1, false); err != nil {
 		return err
 	}
 
-	if err := validate.MaximumInt("policy_type_id", "path", int64(o.PolicyTypeID), 2.147483647e+09, false); err != nil {
+	if err := validate.MaximumInt("policy_type_id", "path", o.PolicyTypeID, 2.147483647e+09, false); err != nil {
 		return err
 	}
